# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fnashha Home Services Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests/{requestId}/offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List offers for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Submit an offer
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /requests/{requestId}/offers/{offerId}:
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /requests/{requestId}/offers/{offerId}/select:
    post:
      operationId: selectOffer
      tags: [offers]
      summary: Customer selects a technician offer
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK

  # ─── MESSAGES ────────────────────────────────────────────────────────────────
...
  /points/transactions:
    get:
      operationId: listPointTransactions
      tags: [points]
      summary: List point transactions
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PointTransaction"
  /points/add:
    post:
      operationId: addPoints
      tags: [points]
      summary: Add points to technician (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"
  /points/deduct:
    post:
      operationId: deductPoints
      tags: [points]
      summary: Deduct points from technician (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PointsInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsBalance"

  # ─── COMMISSIONS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, mobile, role, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [customer, technician, admin, super_admin]
        status:
          type: string
          enum: [active, pending, suspended, banned, rejected]
        profileImage:
          type: ["string", "null"]
        jobTitle:
          type: ["string", "null"]
        createdAt:
          type: string
        technicianProfile:
          $ref: "#/components/schemas/TechnicianProfile"
...
    TechnicianProfile:
      type: object
      required: [id, userId, nationalId, approvalStatus]
      properties:
        id:
          type: integer
        userId:
          type: integer
        nationalId:
          type: string
        personalPhoto:
          type: ["string", "null"]
        nationalIdFront:
          type: ["string", "null"]
        nationalIdBack:
          type: ["string", "null"]
        approvalStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        pointsBalance:
          type: integer
        averageRating:
          type: number
        reviewCount:
          type: integer
        completedJobs:
          type: integer
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        areas:
          type: array
          items:
            $ref: "#/components/schemas/Area"
        primaryArea:
          $ref: "#/components/schemas/Area"
...
    Service:
      type: object
      required: [id, name, isActive, displayOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        icon:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        isActive:
          type: boolean
        displayOrder:
          type: integer
...
    Governorate:
      type: object
      required: [id, name, nameAr, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        isActive:
          type: boolean
...
    Area:
      type: object
      required: [id, name, nameAr, governorateId, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        governorateId:
          type: integer
        governorate:
          $ref: "#/components/schemas/Governorate"
        isActive:
          type: boolean
...
    Offer:
      type: object
      required: [id, requestId, technicianId, price, status, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        technicianId:
          type: integer
        technician:
          $ref: "#/components/schemas/User"
        price:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, selected, rejected, withdrawn]
        createdAt:
          type: string
    OfferInput:
      type: object
      required: [price]
      properties:
        price:
          type: number
        notes:
          type: string
    OfferUpdate:
      type: object
      properties:
        price:
          type: number
        notes:
          type: string

    # ─── MESSAGE ───────────────────────────────────────────────────────────────
...
    PointsBalance:
      type: object
      required: [technicianId, balance]
      properties:
        technicianId:
          type: integer
        balance:
          type: integer
    PointTransaction:
      type: object
      required: [id, technicianId, amount, type, description, balanceAfter, createdAt]
      properties:
        id:
          type: integer
        technicianId:
          type: integer
        amount:
          type: integer
        type:
          type: string
          enum: [credit, debit, commission]
        description:
          type: string
        balanceAfter:
          type: integer
        requestId:
          type: ["integer", "null"]
        createdAt:
          type: string
    PointsInput:
      type: object
      required: [technicianId, amount, description]
      properties:
        technicianId:
          type: integer
        amount:
          type: integer
        description:
          type: string

    # ─── COMMISSION ────────────────────────────────────────────────────────────
...